requirements.txt
requests
googletrans==4.0.0-rc1
PyGithub
python-dotenv
